Make `cltv_expiry_delta` `u16`
The BOLT 7 wire format defines `cltv_expiry_delta` as a 2-byte field,
and LDK uses u16 for it everywhere (`RouteHintHop`, `ChannelUpdateInfo`,
`UnsignedChannelUpdate`). Align the LSPS2 types accordingly.

serde_json will reject values exceeding `u16::MAX` during
deserialization, so a counterparty sending an out-of-range value is
handled gracefully.

+++ lightning-liquidity/src/lsps2/event.rs
@@ -73,7 +73,7 @@ pub enum LSPS2ClientEvent {
 		/// The intercept short channel id to use in the route hint.
 		intercept_scid: u64,
 		/// The `cltv_expiry_delta` to use in the route hint.
-		cltv_expiry_delta: u32,
+		cltv_expiry_delta: u16,
 		/// The initial payment size you specified.
 		payment_size_msat: Option<u64>,
 	},
